+++ DB script/ShoppingCardDB/Table Script/Shopping Cart DB Script.sql
@@ -16,10 +16,10 @@ USE `shoppingcartdb` ;
 -- -----------------------------------------------------
 DROP TABLE IF EXISTS `shoppingcartdb`.`tb_shopping_cart` ;
 CREATE TABLE IF NOT EXISTS `shoppingcartdb`.`tb_shopping_cart` (
-  `ID` INT NOT NULL AUTO_INCREMENT,
+  `ID` BIGINT NOT NULL AUTO_INCREMENT,
   `PRICE` DECIMAL(6,0) NULL DEFAULT NULL,
   `DISCOUNT` DECIMAL(3,0) NULL DEFAULT NULL,
-  `TB_USER_ID` INT NOT NULL,
+  `TB_USER_ID` BIGINT NOT NULL,
   PRIMARY KEY (`ID`))
 ENGINE = InnoDB
 DEFAULT CHARACTER SET = utf8mb4
@@ -30,9 +30,9 @@ COLLATE = utf8mb4_0900_ai_ci;
 -- -----------------------------------------------------
 DROP TABLE IF EXISTS `shoppingcartdb`.`tb_shopping_cart_has_tb_product` ;
 CREATE TABLE IF NOT EXISTS `shoppingcartdb`.`tb_shopping_cart_has_tb_product` (
-  `ID` INT NOT NULL AUTO_INCREMENT,
-  `TB_SHOPPING_CART_ID` INT NOT NULL,
-  `TB_PRODUCT_ID` INT NOT NULL,
+  `ID` BIGINT NOT NULL AUTO_INCREMENT,
+  `TB_SHOPPING_CART_ID` BIGINT NOT NULL,
+  `TB_PRODUCT_ID` BIGINT NOT NULL,
   `Quantity` INT NOT NULL,
   PRIMARY KEY (`ID`),
   INDEX `TB_SHOPPING_CART_ID_idx` (`TB_SHOPPING_CART_ID` ASC) VISIBLE,
